requirements.txt
Flask==2.3.2
numpy==1.24.3
Pillow==10.2.0
tensorflow==2.12.0
Werkzeug==2.3.7
